requirements.txt
openai
python-dotenv
requests
beautifulsoup4
tweepy
pytrends
apscheduler
selenium
scrapy
spacy
transformers
python-namecheap
playwright
